chore: Move Zutil for common-service-core
