+++ common-service-core/src/main/java/org/zowe/apiml/product/zos/ZUtil.java
@@ -8,18 +8,15 @@
  * Copyright Contributors to the Zowe Project.
  */
 
-package org.zowe.apiml.eurekaservice.client.impl;
+package org.zowe.apiml.product.zos;
 
 import java.io.OutputStream;
 import java.io.PrintStream;
 import java.util.Properties;
 
-/**
+/*
  * For API documentation see https://www.ibm.com/support/knowledgecenter/SSYKE2_8.0.0/com.ibm.java.zsecurity.api.80.doc/com.ibm.jzos/com/ibm/jzos/ZUtil.html
- *
- * @deprecated Use #{org.zowe.apiml.product.zos.ZUtil} from apiml-utility
  */
-@Deprecated
 public interface ZUtil {
 
     String[] environ();
